services/action-gateway/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
discord.py==2.5.2
pyyaml==6.0.2
aiosqlite==0.21.0
